functions/deleteAccount/requirements.txt
appwrite
python-dotenv